requirements.txt
alembic==1.7.5
attrs==21.2.0
autopage==0.4.0
backcall==0.2.0
certifi==2020.6.20
cliff==3.10.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.3.2
colorlog==6.6.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
greenlet==1.1.2
gym==0.17.3
gym-climate==0.0.1
gym-conservation==0.0.6
gym-fishing==0.1.0
importlib-metadata==4.8.2
importlib-resources==5.4.0
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kaleido==0.2.1
kiwisolver==1.2.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.3.2
numpy==1.19.2
opencv-python==4.5.5.62
optuna==2.10.0
packaging==21.3
pandas==1.1.3
parso==0.7.1
pbr==5.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==5.4.0
prettytable==2.4.0
prompt-toolkit==3.0.7
ptyprocess==0.6.0
pyglet==1.5.0
Pygments==2.7.1
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.1
python-dotenv==0.19.2
pytz==2020.1
PyYAML==6.0
sb3-contrib==1.4.0
scipy==1.5.2
six==1.15.0
SQLAlchemy==1.4.27
stable-baselines3==1.4.0
stevedore==3.5.0
tenacity==8.0.1
torch==1.10.0
tqdm==4.62.3
traitlets==5.0.4
typing_extensions==4.0.0
wcwidth==0.2.5
zipp==3.6.0